0.1.5/ OrderDetails modal window refactoring according to the task
